fix: rename ui-> prebuiltui to avoid import conflicts
